# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.CreateExpensePlanRecordRequest:
    properties:
      amount_paid:
        type: integer
      expense_plan_id:
        type: integer
      paid_date:
        type: string
      payment_date:
        type: string
    type: object
  handler.CreateExpensePlanRecordResponse:
    properties:
      expense_plan_record_id:
        type: integer
    type: object
  handler.CreateExpensePlanRequest:
    properties:
      amount_planned:
        type: integer
      category:
        $ref: '#/definitions/query.ExpensePlanCategory'
      recurrency_interval:
        type: integer
      recurrency_type:
        $ref: '#/definitions/query.RecurrencyType'
      title:
        type: string
    required:
    - amount_planned
    - category
    - title
    type: object
  handler.CreateExpensePlanResponse:
    properties:
      expense_plan_id:
        type: integer
    type: object
  handler.DeleteExpensePlanRecordsRequest:
    properties:
      expense_plan_record_id:
        type: integer
    required:
    - expense_plan_record_id
    type: object
  handler.DeleteExpensePlanRecordsResponse:
    type: object
  handler.DeleteExpensePlanRequest:
    properties:
      expense_plan_id:
        type: integer
    required:
    - expense_plan_id
    type: object
  handler.DeleteExpensePlanResponse:
    type: object
  handler.GetExpensePlanRecordRequest:
    properties:
      expense_plan_id:
        type: integer
    required:
    - expense_plan_id
    type: object
  handler.GetExpensePlanRecordResponse:
    properties:
      amount_paid:
        type: integer
      created_at:
        type: string
      expense_plan:
        $ref: '#/definitions/handler.GetExpensePlanRecordResponse_ExpensePlan'
      expense_plan_id:
        type: integer
      expense_plan_record_id:
        type: integer
      expense_plan_sequence:
        type: integer
      paid_date:
        type: string
      payment_date:
        type: string
      updated_at:
        type: string
    type: object
  handler.GetExpensePlanRecordResponse_ExpensePlan:
    properties:
      amount_planned:
        type: integer
      category:
        type: string
      expense_plan_id:
        type: integer
      recurrency_interval:
        type: integer
      recurrency_type:
        $ref: '#/definitions/query.RecurrencyType'
      title:
        type: string
    type: object
  handler.GetExpensePlanRequest:
    properties:
      expense_plan_id:
        type: integer
    required:
    - expense_plan_id
    type: object
  handler.GetExpensePlanResponse:
    properties:
      amount_planned:
        type: integer
      category:
        $ref: '#/definitions/query.ExpensePlanCategory'
      created_at:
        type: string
      expense_plan_id:
        type: integer
      first_paid_date:
        type: string
      last_amount_spent:
        type: integer
      last_paid_date:
        type: string
      paid_count:
        type: integer
      recurrency_interval:
        type: integer
      recurrency_type:
        $ref: '#/definitions/query.RecurrencyType'
      title:
        type: string
      updated_at:
        type: string
    type: object
  handler.ListExpensePlanRecordRequest:
    type: object
  handler.ListExpensePlanRecordResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handler.ListExpensePlanRecordResponse_ListExpensePlanRecord'
        type: array
    type: object
  handler.ListExpensePlanRecordResponse_ListExpensePlan:
    properties:
      amount_planned:
        type: integer
      category:
        type: string
      expense_plan_id:
        type: integer
      recurrency_interval:
        type: integer
      recurrency_type:
        $ref: '#/definitions/query.RecurrencyType'
      title:
        type: string
    type: object
  handler.ListExpensePlanRecordResponse_ListExpensePlanRecord:
    properties:
      amount_paid:
        type: integer
      created_at:
        type: string
      expense_plan:
        $ref: '#/definitions/handler.ListExpensePlanRecordResponse_ListExpensePlan'
      expense_plan_id:
        type: integer
      expense_plan_record_id:
        type: integer
      expense_plan_sequence:
        type: integer
      paid_date:
        type: string
      payment_date:
        type: string
      updated_at:
        type: string
    type: object
  handler.ListExpensePlanRequest:
    type: object
  handler.ListExpensePlanResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handler.ListExpensePlanResponse_ListExpensePlan'
        type: array
    type: object
  handler.ListExpensePlanResponse_ListExpensePlan:
    properties:
      amount_planned:
        type: integer
      category:
        $ref: '#/definitions/query.ExpensePlanCategory'
      created_at:
        type: string
      expense_plan_id:
        type: integer
      first_paid_date:
        type: string
      last_amount_spent:
        type: integer
      last_paid_date:
        type: string
      paid_count:
        type: integer
      recurrency_interval:
        type: integer
      recurrency_type:
        $ref: '#/definitions/query.RecurrencyType'
      title:
        type: string
      updated_at:
        type: string
    type: object
  handler.UpdateExpensePlanRecordRequest:
    properties:
      amount_paid:
        type: integer
      expense_plan_id:
        type: integer
      expense_plan_record_id:
        type: integer
      paid_date:
        type: string
      payment_date:
        type: string
    type: object
  handler.UpdateExpensePlanRecordResponse:
    properties:
      expense_plan_record_id:
        type: integer
    type: object
  handler.UpdateExpensePlanRequest:
    properties:
      amount_planned:
        type: integer
      category:
        $ref: '#/definitions/query.ExpensePlanCategory'
      expense_plan_id:
        type: integer
      recurrency_interval:
        type: integer
      recurrency_type:
        $ref: '#/definitions/query.RecurrencyType'
      title:
        type: string
    required:
    - amount_planned
    - category
    - expense_plan_id
    - title
    type: object
  handler.UpdateExpensePlanResponse:
    properties:
      expense_plan_id:
        type: integer
    type: object
  query.ExpensePlanCategory:
    enum:
    - FOOD
    - TRANSPORT
    - PROPERTY
    - TAX
    - ENTERTAINMENT
    - OTHER
    type: string
    x-enum-varnames:
    - ExpensePlanCategoryFOOD
    - ExpensePlanCategoryTRANSPORT
    - ExpensePlanCategoryPROPERTY
    - ExpensePlanCategoryTAX
    - ExpensePlanCategoryENTERTAINMENT
    - ExpensePlanCategoryOTHER
  query.RecurrencyType:
    enum:
    - MONTHLY
    - YEARLY
    type: string
    x-enum-varnames:
    - RecurrencyTypeMONTHLY
    - RecurrencyTypeYEARLY
host: localhost:8080
info:
  contact: {}
  description: This is a sample API using Swagger.
  title: My API
  version: "1.0"
paths:
  /expense-plan-record/create:
    post:
      consumes:
      - application/json
      description: Create expense-plan-record item
      parameters:
      - description: CreateExpensePlanRecordRequest
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.CreateExpensePlanRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CreateExpensePlanRecordResponse'
      summary: Create expense-plan-record item
      tags:
      - ExpensePlanRecord
  /expense-plan-record/delete:
    post:
      consumes:
      - application/json
      description: Delete expense-plan-record item
      parameters:
      - description: DeleteExpensePlanRecordsRequest
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.DeleteExpensePlanRecordsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.DeleteExpensePlanRecordsResponse'
      summary: Delete expense-plan-record item
      tags:
      - ExpensePlanRecords
  /expense-plan-record/get:
    post:
      consumes:
      - application/json
      description: Get all expense-plan-record items (using cursor-based pagination)
      parameters:
      - description: GetExpensePlanRecordRequest
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.GetExpensePlanRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetExpensePlanRecordResponse'
      summary: Get all expense-plan-record items
      tags:
      - ExpensePlanRecord
  /expense-plan-record/list:
    post:
      consumes:
      - application/json
      description: List all expense-plan-record items (using cursor-based pagination)
      parameters:
      - description: ListExpensePlanRecordRequest
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.ListExpensePlanRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListExpensePlanRecordResponse'
      summary: List all expense-plan-record items
      tags:
      - ExpensePlanRecord
  /expense-plan-record/update:
    post:
      consumes:
      - application/json
      description: Update expense-plan-record item
      parameters:
      - description: UpdateExpensePlanRecordRequest
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.UpdateExpensePlanRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UpdateExpensePlanRecordResponse'
      summary: Update expense-plan-record item
      tags:
      - ExpensePlanRecord
  /expense-plan/create:
    post:
      consumes:
      - application/json
      description: Create expense-plan item
      parameters:
      - description: CreateExpensePlanRequest
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.CreateExpensePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CreateExpensePlanResponse'
      summary: Create expense-plan item
      tags:
      - ExpensePlan
  /expense-plan/delete:
    post:
      consumes:
      - application/json
      description: Delete expense-plan item
      parameters:
      - description: DeleteExpensePlanRequest
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.DeleteExpensePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.DeleteExpensePlanResponse'
      summary: Delete expense-plan item
      tags:
      - ExpensePlan
  /expense-plan/get:
    post:
      consumes:
      - application/json
      description: Get all expense-plan items (using cursor-based pagination)
      parameters:
      - description: GetExpensePlanRequest
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.GetExpensePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetExpensePlanResponse'
      summary: Get all expense-plan items
      tags:
      - ExpensePlan
  /expense-plan/list:
    post:
      consumes:
      - application/json
      description: List all expense-plan items (using cursor-based pagination)
      parameters:
      - description: ListExpensePlanRequest
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.ListExpensePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListExpensePlanResponse'
      summary: List all expense-plan items
      tags:
      - ExpensePlan
  /expense-plan/update:
    post:
      consumes:
      - application/json
      description: Update expense-plan item
      parameters:
      - description: UpdateExpensePlanRequest
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.UpdateExpensePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UpdateExpensePlanResponse'
      summary: Update expense-plan item
      tags:
      - ExpensePlan
swagger: "2.0"